requirements.txt
aioredis==1.3.1
amqp==5.0.6
asgiref==3.3.4
async-timeout==3.0.1
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
billiard==3.6.4.0
celery==5.0.5
cffi==1.14.5
channels==3.0.3
channels-redis==3.2.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
constantly==15.1.0
coverage==5.5
cryptography==3.4.7
daphne==3.0.2
Django==3.1.7
django-autoslug==1.9.8
django-autoslug-field==0.2.3
django-celery-beat==2.2.0
django-celery-results==2.0.1
django-cors-headers==3.7.0
django-coupon-management==0.3.1
django-debug-toolbar==3.2.1
django-environ==0.4.5
django-filter==2.4.0
django-model-utils==4.1.1
django-notifications==0.1.dev0
django-notifications-hq==1.6.0
django-notifications-rest==0.0.1
django-phonenumber-field==5.2.0
django-pytest==0.2.0
django-timezone-field==4.1.2
djangorestframework==3.12.4
hiredis==2.0.0
hyperlink==21.0.0
idna==3.2
incremental==21.3.0
iniconfig==1.1.1
jsonfield==3.1.0
kombu==5.1.0
msgpack==1.0.2
packaging==20.9
phonenumbers==8.12.27
phonenumberslite==8.12.27
Pillow==7.0.0
pluggy==0.13.1
prompt-toolkit==3.0.19
psycopg2-binary==2.8.6
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyOpenSSL==20.0.1
pyparsing==2.4.7
pytest==6.2.4
pytest-cov==2.12.1
pytest-django==4.3.0
pytest-sugar==0.9.4
python-crontab==2.5.1
python-dateutil==2.8.1
pytz==2021.1
redis==3.5.3
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.1
swapper==1.1.2.post1
termcolor==1.1.0
toml==0.10.2
Twisted==21.2.0
txaio==21.2.1
vine==5.0.0
wcwidth==0.2.5
zope.interface==5.4.0
